resolve  incomplete type
