--- Examples/minimum_user_for_s2e/src/src_user/Settings/TlmCmd/Ccsds/space_packet_len_params.h
@@ -1,14 +1,0 @@
-/**
- * @file
- * @brief CmdSpacePacket のユーザー設定
- */
-#ifndef SPACE_PACKET_LEN_PARAMS_H_
-#define SPACE_PACKET_LEN_PARAMS_H_
-
-#undef TSP_MAX_LEN
-#undef CSP_MAX_LEN
-
-#define TSP_MAX_LEN  (432)
-#define CSP_MAX_LEN  (128)
-
-#endif
