+++ Examples/minimum_user_for_s2e/src/src_user/Settings/TlmCmd/Ccsds/space_packet_typedef_params.h
@@ -1,0 +1,14 @@
+/**
+ * @file
+ * @brief SpacePacketの パケット型に関するユーザー設定
+ */
+#ifndef SPACE_PACKET_TYPEDEF_PARAMS_H_
+#define SPACE_PACKET_TYPEDEF_PARAMS_H_
+
+#undef TSP_MAX_LEN
+#undef CSP_MAX_LEN
+
+#define TSP_MAX_LEN  (432)
+#define CSP_MAX_LEN  (128)
+
+#endif
